google_sheets.py
import gspread
from google.oauth2.service_account import Credentials
import asyncio
from datetime import datetime
import logging

class GoogleSheetsManager:
    def __init__(self):
        # Настройка доступа к Google Sheets API
        # Для тестирования используем публичный доступ
        self.sheet_url = "https://docs.google.com/spreadsheets/d/1pIWAwVRL3YC603p_8JxF5827AbmIYCRddsoXZpQshVI/edit?usp=sharing"
        self.sheet_id = "1pIWAwVRL3YC603p_8JxF5827AbmIYCRddsoXZpQshVI"
        
    async def add_order(self, order_data: dict):
        """Добавление заказа в Google Sheets"""
        try:
            # Для демонстрации просто логируем данные
            # В реальном проекте здесь будет код для работы с Google Sheets API
            logging.info(f"Заказ добавлен в таблицу: {order_data}")
            
            # Имитация асинхронной операции
            await asyncio.sleep(0.1)
            
            # Здесь должен быть код для записи в Google Sheets:
            # 1. Аутентификация через service account
            # 2. Открытие таблицы
            # 3. Добавление строки с данными заказа
            
            return True
            
        except Exception as e:
            logging.error(f"Ошибка при добавлении заказа в таблицу: {e}")
            return False
    
    def get_orders_count(self):
        """Получение количества заказов"""
        # Заглушка для демонстрации
        return 0
