requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0
